Created SearchResult component to appear on search page and added addBoook function to
For every search result a SearchResult component will be created. Also added the addBook function in the app component. It gets passed to the Search component and then to the SearchResult allowing the shelf value of the search result to be changed and added both to state and the BooksAPI